backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
SQLAlchemy==2.0.43
PyMySQL==1.1.1
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
PyJWT==2.10.1
bcrypt==4.0.1
scikit-learn==1.5.2
python-multipart==0.0.9
opencv-python-headless>=4.8.0
Pillow>=10.0.0
tensorflow>=2.16.0
tf-keras>=2.16.0
deepface>=0.0.92
easyocr>=1.7.0
